refactoring

+++ src/main/java/com/training/salestaxes/calculation/TaxCalculation.java
@@ -1,8 +1,9 @@
 package com.training.salestaxes.calculation;
 
+import java.math.BigDecimal;
 import com.training.salestaxes.items.Product;
 
 public interface TaxCalculation
 {
-	double calculateTaxesOn(Product product);
+	BigDecimal calculateTaxesOn(Product product);
 }
